requirements.txt
Flask==2.1.2
Flask_Cors==3.0.10
matplotlib==3.5.2
numpy==1.22.4
pandas==1.4.3
pymongo==4.2.0
python_bcrypt==0.3.2